+++ pkgs/google_cloud_auth/lib/google_cloud_auth.dart
@@ -15,4 +15,7 @@
 /// Authentication and credential management for Google Cloud.
 library;
 
+export 'src/compute_engine_credentials.dart';
+export 'src/credential_exception.dart';
 export 'src/service_account_credentials.dart';
+export 'src/service_account_signer.dart';